v0.1.0 - see changelog for details

+++ scripts/mongo.js
@@ -1,5 +1,19 @@
-db.lookup.drop();
-db.county.drop();
+print('Dropping collections');
+
+var drop1 = db.lookup.drop();
+
+if (!drop1) {
+    exit();
+};
+
+var drop2 = db.county.drop();
+
+if (!drop2) {
+    exit();
+};
+
+print('Inserting data for tests');
+
 db.lookup.insert([{
     recType: 'bankCodeType',
     luValue: 'bsb',
@@ -10,6 +24,7 @@ db.lookup.insert([{
     luValue: 'iban',
     valueText: 'IBAN'
 }]);
+
 db.county.insert([{
     _id: ObjectId('55e071b8aa1643a3971c430f'),
     county: 'Berkshire',
@@ -19,3 +34,5 @@ db.county.insert([{
     county: 'Bedfordshire',
     country: 'England'
 }]);
+
+print('Finished');
